feat: enhance security, caching, and hash comparison automation
- Add comprehensive API security regression tests covering token masking, TLS flags, request injection, and rate limiter bypass
- Harden security gateway with enhanced validation and monitoring
- Update security documentation with concise best practices

- Implement automatic hash comparison step generation for checksum verification
- Add `examples/json/hash_compare.json` demonstrating multi-command expansion
- Enhance suggestion system with multi-command parser integration

- Improve LRU cache with batch eviction and pathological insert detection
- Optimize blueprint registry with disk caching and progress logging
- Add project ID validation enforcing signed 64-bit limits

- Enforce minimum pytest collection count in CI to maintain coverage integrity
- Add development tools directory for analysis and benchmarking
- Update example workflows with proper Python path handling
- Remove deprecated `robot.utils` compatibility shim

- Update CHANGELOG with detailed 0.1.2 release notes
- Improve security guidelines with specific monitoring recommendations
- Clean up documentation language removing marketing fluff

- Expand test coverage with API security and JSON cache serialization tests
- Add hash comparison and medallion configuration test modules
- Fix test isolation issues with thread-local context patterns

+++ src/importobot/caching/base.py
@@ -24,7 +24,7 @@ class CacheConfig:
     """Unified cache configuration."""
 
     max_size: int = 1000
-    ttl_seconds: int | None = None
+    ttl_seconds: float | None = None
     eviction_policy: EvictionPolicy = EvictionPolicy.LRU
     max_content_size_bytes: int = 50000
     enable_telemetry: bool = True
